fixed a warning
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@15556 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/tiff/tif_print.c
@@ -460,7 +460,7 @@ _TIFFprintAscii(FILE* fd, const char* cp)
 	for (; *cp != '\0'; cp++) {
 		const char* tp;
 
-		if (isprint(*cp)) {
+		if (isprint((int)*cp)) {
 			fputc(*cp, fd);
 			continue;
 		}
